Some light refactoring
Split code into multiple files and made the chunk_add/subtract function
signatures more consistent with each other.
